chore(ui): move git status tree into separate module
